// src/types/public/Dentists.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ClinicsId } from './Clinics';
import { DentistClinic, type DentistDegrees } from './DentistDegrees';
import DentistBranches, { type DentistBranchesId } from './DentistBranches';

/** Identifier type for public.dentists */
export type DentistsId = number & { __brand: 'DentistsId' };

/** Represents the table public.dentists */
export default interface Dentists {
  id: DentistsId;

  clinic_id: ClinicsId;

  clinic_name: string

  email: string | null;

  name: string;

  last_name: string;

  phone: string | null;

  identity_no: string | null;

  spoken_languages: unknown | null;

  keywords: unknown | null;

  seo_settings: unknown | null;

  degrees: DentistDegrees;

  branch: any | null;

  url: string | null;

  is_active: boolean | null;

  tr: string

  cover_image: string

  created_at: Date;

  updated_at: Date;
  clinics: DentistClinic;
  address: string;
  dentist_branches: DentistBranches;
  city: string;
  review_rating: string;
  review_count: number;

}

/** Represents the initializer for the table public.dentists */
export interface DentistsInitializer {
  /** Default value: nextval('dentists_id_seq'::regclass) */
  id?: DentistsId;

  clinic_id: ClinicsId;

  email?: string | null;

  name: string;

  last_name: string;

  phone?: string | null;

  identity_no?: string | null;

  /** Default value: '["tr"]'::jsonb */
  spoken_languages?: unknown | null;

  /** Default value: '[]'::jsonb */
  keywords?: unknown | null;

  /** Default value: '{}'::jsonb */
  seo_settings?: unknown | null;

  degrees?: DentistDegrees | null;

  branch_id?: DentistBranchesId | null;

  url?: string | null;

  /** Default value: false */
  is_active?: boolean | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;
}

/** Represents the mutator for the table public.dentists */
export interface DentistsMutator {
  id?: DentistsId;

  clinic_id?: ClinicsId;

  email?: string | null;

  name?: string;

  last_name?: string;

  phone?: string | null;

  identity_no?: string | null;

  spoken_languages?: unknown | null;

  keywords?: unknown | null;

  seo_settings?: unknown | null;

  degrees?: DentistDegrees | null;

  branch_id?: DentistBranchesId | null;

  url?: string | null;

  is_active?: boolean | null;

  created_at?: Date;

  updated_at?: Date;
}